LayoutTests/imported/w3c/web-platform-tests/cookie-store/cookieStore_subscribe_arguments.https.any.serviceworker-expected.txt


PASS cookieStore.subscribe without url in option
PASS cookieStore.subscribe with invalid url path in option
PASS cookieStore.subscribe is idempotent
PASS CookieStore.unsubscribe is idempotent
